Make `SocInfo` static
The idea of using `Cow<'a, str>` everywhere was that Serde would deserialise the borrowed variant for most strings, and then if any strings used escape characters they'd be owned instead. But I never actually checked if it did this, and it doesn't: it always uses the owned version, presumably so that `Cow` can implement `DeserializeOwned`.

So, now it just uses `String`s instead.